fix(node): forward proposal ancestors if missing (#408)
This patch addresses an edge case where notarized (but not yet
finalized) blocks are not forwarded to the execution layer, stopping the
chain from proposing or validating new blocks.

This fix consists in the execution driver always attempting to backfill
the parent block of a given proposal or digest. It does so by starting
at the parent and walking its ancestors all the way until reaching the
latest block available on the execution layer.

As part of the patch backfilling at node startup and during steady state
operation was unified.

This happens under the following scenario:

1. all validators are at a finalized height H and notarized height H+1
2. all validators' execution layers have finalized = H, head = H+1
3. the validators have not witnessed a quorum of finalization receipts
4. all validators are stopped at once
5. upon restart, the validators' consensus layer is caught up to
finalized H and notarized H+1
6. the execution layers are at finalized = head = H
7. the EL has not persisted H+1 since it only persists up to the
finalized head
8. the consensus layers attempt to transition to round H+2 with parent
H+1
9. none of the validators can propose nor verify
10. simplex does not guarantee a direct finalization of H+1, but also no
indirect finalization occurs because the chain does not propose new
blocks
11. the execution layer will never see block H+1
12. the network is stalled