feat!(precompiles): migrate TIP403 registry (#803)
since the `contract` macro was getting exponentially larger and more
complex due to having to distinguish between user-defined structs (not
packable in the layout) and enums (packable, as they are equivalent to
u8s), i ended up not supporting them in the `contract` macro.

Instead, we define them as u8s, and leverage the sol! macro's
`TryInto<u8>` impl, as well as their `as u8` capabilities.